backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pyjwt>=2.3.0
python-dotenv>=0.19.0
pydantic>=1.8.0
stripe>=2.60.0
llama-index>=0.9.0
docker>=5.0.0
python-multipart>=0.0.5